Remove Login component and inline in the view.
