+++ food-things-frontend/tests/unit/views/Login.spec.js
@@ -1,17 +1,19 @@
 import { shallowMount } from '@vue/test-utils'
-import Login from '@/components/Login'
-import registerGlobals from './global/'
+import Login from '@/views/Login'
+import registerGlobals from '../components/global'
+import $store from '@/store'
 
 registerGlobals()
 
 describe('Login', () => {
   it('Should render a div with no classes.', () => {
-    const wrapper = shallowMount(Login)
+    const wrapper = shallowMount(Login, { mocks: { $store } })
+
     expect(wrapper.is('div')).toBe(true)
   })
 
   it('Should render three inputs of correct types.', () => {
-    const wrapper = shallowMount(Login)
+    const wrapper = shallowMount(Login, { mocks: { $store } })
 
     expect(wrapper.find('[type="text"]').exists()).toBe(true)
     expect(wrapper.find('[type="password"]').exists()).toBe(true)
@@ -19,7 +21,7 @@ describe('Login', () => {
   })
 
   it('Should render a button with the type of success.', () => {
-    const wrapper = shallowMount(Login)
+    const wrapper = shallowMount(Login, { mocks: { $store } })
     expect(wrapper.find('[type="success"]').exists()).toBe(true)
   })
 
@@ -29,6 +31,9 @@ describe('Login', () => {
         return {
           submitting: true,
         }
+      },
+      mocks: {
+        $store,
       },
     })
     expect(wrapper.attributes()).toEqual({ fullscreen: 'true' })
